drivers: Split stm32l4 qspi driver to mem-map & flash_dev files
